Extract rest function
