+++ s2core/src/main/scala/com/kakao/s2graph/core/mysqls/Model.scala
@@ -54,6 +54,16 @@ object Model {
       res
     }
   }
+
+  def loadCache() = {
+    Service.findAll()
+    ServiceColumn.findAll()
+    Label.findAll()
+    LabelMeta.findAll()
+    LabelIndex.findAll()
+    ColumnMeta.findAll()
+  }
+
 }
 
 trait Model[V] extends SQLSyntaxSupport[V] {
