+++ s2rest_play/app/controllers/AdminController.scala
@@ -2,6 +2,7 @@ package controllers
 
 import com.kakao.s2graph.core._
 import com.kakao.s2graph.core.mysqls._
+import com.kakao.s2graph.core.rest.RequestParser
 import com.kakao.s2graph.core.utils.logger
 import play.api.mvc
 import play.api.mvc.{Action, Controller}
