+++ s2rest_play/app/controllers/QueryController.scala
@@ -4,6 +4,7 @@ package controllers
 import com.kakao.s2graph.core.GraphExceptions.BadQueryException
 import com.kakao.s2graph.core._
 import com.kakao.s2graph.core.mysqls._
+import com.kakao.s2graph.core.rest.RequestParser
 import com.kakao.s2graph.core.types.{LabelWithDirection, VertexId}
 import com.kakao.s2graph.core.utils.logger
 import config.Config
